svc-admin/requirements.txt
# Core Framework
fastapi==0.115.9
uvicorn[standard]==0.34.3

# Database
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20

# HTTP Client
httpx==0.28.1

# Data Validation & Settings
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0

# Email Services
sendgrid==6.12.4

# Validation
email-validator==2.3.0

# PDF Generation
reportlab==4.4.0

# Type Extensions
typing-extensions==4.13.2
